Setting all floats now to doubles. all tests pass

+++ sofia/cuda/histogram.h
@@ -1,1 +1,1 @@
-void histogram(float* ary, double* bins, int* histo, int ary_len, int bin_len, int histo_len);
+void histogram(double* ary, double* bins, int* histo, int ary_len, int bin_len, int histo_len);
